+++ examples/sandbox/src/main.rs
@@ -15,7 +15,8 @@ impl MyApp {
     pub fn new(_cc: &eframe::CreationContext) -> Self {
         Self {
             file_dialog: FileDialog::new()
-                .anchor(egui::Align2::CENTER_CENTER, egui::Vec2::new(0.0, 0.0)),
+                .anchor(egui::Align2::CENTER_CENTER, egui::Vec2::new(0.0, 0.0))
+                .id("egui_file_dialog"),
 
             selected_directory: None,
             selected_file: None,